Frontend Refactor
Rename folders, files, variables and constants to meet coding standards
