--- Frontend/infosafe_frontend/src/Styling/EditUser.css
@@ -1,164 +1,0 @@
-.editUserOverlay {
-    width: 60vw;
-    height: 65vh;
-    flex-shrink: 0;
-    border-radius: 30px;
-    background: rgba(173, 173, 173, 0.69);
-    backdrop-filter: blur(20px);
-    margin: 0 auto;
-}
-.border {
-    border: 1px solid #ffffff;
-    width: 56vw;
-    margin: 0 auto;
-    height: 57vh;
-    position: relative;
-    top: 4vh;
-}
-.backButton {
-    background-color: rgba(233, 0, 0, 1);
-    border: none;
-    position: relative;
-    height: 4vh;
-    width: 2.5vw;
-    align-content: center;
-    border-radius: 5px;
-    top: 2vh;
-    left: 1vw;
-    color: #ffffff;
-}
-.backIcon {
-    color: #ffffff;
-    margin: 0 auto;
-    height: 3vh;
-    width: 1.3vw;
-}
-.editUserTitle {
-    width: 60%;
-    justify-content: center;
-    color: #000000;
-    text-align: center;
-    text-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-    font-size: 35px;
-    font-family: 'Roboto', sans-serif;
-    text-transform: uppercase;
-    position: relative;
-    top: -2vh;
-    left: 20%;
-}
-.nameEdit {
-    margin: 0 auto;
-    justify-content: center;
-    align-content: center;
-    color: #000;
-    text-align: center;
-    font-size: 21px;
-    text-transform: uppercase;
-    font-family: 'Roboto', sans-serif;
-    position: relative;
-    top: 1vh;
-}
-.editNameInput {
-    font-size: 20px;
-    width: 50%;
-    text-align: center;
-    border: none;
-    height: 4vh;
-}
-.surnameEdit {
-    margin: 0 auto;
-    justify-content: center;
-    align-content: center;
-    color: #000;
-    text-align: center;
-    font-size: 21px;
-    text-transform: uppercase;
-    font-family: 'Roboto', sans-serif;
-    position: relative;
-    top: 0.5vh;
-    margin-top: 1vh;
-}
-.editSurnameInput {
-    font-size: 20px;
-    width: 50%;
-    text-align: center;
-    border: none;
-    height: 4vh;
-}
-.emailEdit {
-    margin: 0 auto;
-    justify-content: center;
-    align-content: center;
-    color: #000;
-    text-align: center;
-    font-size: 21px;
-    text-transform: uppercase;
-    font-family: 'Roboto', sans-serif;
-    position: relative;
-    top: 0.5vh;
-    margin-top: 1vh;
-}
-.editEmailInput {
-    font-size: 20px;
-    width: 50%;
-    text-align: center;
-    border: none;
-    height: 4vh;
-}
-.roleEdit {
-    margin: 0 auto;
-    justify-content: center;
-    align-content: center;
-    color: #000;
-    text-align: center;
-    font-size: 21px;
-    text-transform: uppercase;
-    font-family: 'Roboto', sans-serif;
-    position: relative;
-    top: 0.5vh;
-    margin-top: 1vh;
-}
-.roleDropdown {
-    text-transform: uppercase;
-    font-size: 18px;
-    width: 50%;
-    margin: 0 auto;
-    text-align: center;
-    border: none;
-    text-align: center;
-    align-content: center;
-    height: 4vh;
-}
-.roleDropdown:after {
-    transform: translateY(50%);
-}
-.FinishButton {
-    width: 15%;
-    height: 5%;
-    background-color: rgba(233, 0, 0, 1);
-    color: #ffffff;
-    text-align: center;
-    font-size: 18px;
-    font-family: 'Roboto', sans-serif;
-    letter-spacing: 0.84px;
-    text-transform: uppercase;
-    position: relative;
-    align-self: center;
-    justify-self: center;
-    margin: 0 auto;
-    top: 5vh;
-    left: 42.5%;
-    border: none;
-}
-.nameTitle {
-    margin-bottom: 0.5vh;
-}
-.surnameTitle {
-    margin-bottom: 0.5vh;
-}
-.emailTitle {
-    margin-bottom: 0.5vh;
-}
-.roleTitle {
-    margin-bottom: 0.5vh;
-}
