+++ frontend/infosafe_frontend/.eslintignore
