+++ frontend/infosafe_frontend/.eslintrc.json
