+++ frontend/infosafe_frontend/.prettierignore
