+++ frontend/infosafe_frontend/package.json
