+++ frontend/infosafe_frontend/public/manifest.json
