+++ frontend/infosafe_frontend/src/components/__tests__/CreateDevicePopup.test.js
