+++ frontend/infosafe_frontend/src/components/__tests__/Home.test.js
