+++ frontend/infosafe_frontend/src/components/__tests__/NavBar.test.js
