+++ frontend/infosafe_frontend/src/components/__tests__/ViewDataScope.test.js
