+++ frontend/infosafe_frontend/src/components/__tests__/ViewDevice.test.js
