+++ frontend/infosafe_frontend/src/components/__tests__/ViewUser.test.js
