+++ frontend/infosafe_frontend/src/setupTests.js
