+++ frontend/infosafe_frontend/src/styling/CreateDataScopePopup.css
