+++ frontend/infosafe_frontend/src/styling/CreateDevicePopup.css
