+++ frontend/infosafe_frontend/src/styling/EditDataScopePopup.css
