+++ frontend/infosafe_frontend/src/styling/EditDevice.css
