+++ frontend/infosafe_frontend/src/styling/Home.css
