+++ frontend/infosafe_frontend/src/styling/NavBar.css
