+++ frontend/infosafe_frontend/src/styling/ViewDataScope.css
