+++ frontend/infosafe_frontend/src/styling/ViewDevice.css
